requirements.txt
discord.py>=2.3.0
Pillow>=9.0.0
pytesseract>=0.3.10
requests>=2.28.0
